add: Sourcing branch files from init.lua
The structure of the rc is that every part of the configuration is
written in a different file that is sourced in the main init.lua
configuration file. There will also be a config file for each plugin.
These files are stored in the nvimrc folder in the runtime. The current
configuration braches are:

- autocmds
- colours
- mappings
- options
- plugins

This list may change in either direction (files added or removed).