edit(refactor): big plugin management refactor
Refactor of packer plugin management...

- Why it needed to be changed?
Because it didn't let me define different parts of my used plugin
ecosystem in different files and the packer spec file was getting over
1000 lines long, which was becoming impractical when edditing plugins.

- Advantages of this restructure:
1) I do not need to define Packer commands with `packadd packer.nvim`
   myself because they are in the packer spec itself.
2) PackerProfile, now functions as ussual
3) PackerCompile now does not break the highlighting (as of now, maybe
   this will break or is broken and funcioning only just now)
4) Allows for dependent plugins to be in one place and therefore the
   file is decoupled into different stages, where different features are
   setup.
5) Do not have to create lua funcions for packer functions in MUtils because
   they are defined in packer itself.

TODO: Need to sort the plugins where they belong and sequence the
loading. There will be a lot, that can be removed when this is done,
because some of the plugin blocks are surely duplicated in the current
setup. <10-03-22, kunzaatko> #