fix(LSP): wrong pattern in on_attach generated autocommand
The pattern from `LspAttach` is the buffer name, but I used the server
name. The server name can be extracted from the client which is then
used to check the name.

TODO: Could be a more general pattern and could use some API to check if
the pattern matches as if the `server_name` was not the exact name but
a pattern similar to autocommand matches.